+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -29,14 +29,14 @@ public class UserController {
 
     @PostMapping
     public @ResponseBody User create(@RequestBody User user) throws ValidateException {
-        validateNewUser(user);
+        userService.validateNewUser(user);
         return userService.createUser(user);
     }
 
 
     @PutMapping
     public @ResponseBody User update(@RequestBody User user) throws ValidateException {
-        validateNewUser(user);
+        userService.validateNewUser(user);
         return userService.updateUser(user);
     }
 
@@ -46,29 +46,6 @@ public class UserController {
     }
 
 
-    public void validateNewUser(User user) throws ValidateException {
-
-        if (user.getName() == null && user.getLogin() == null) {
-            log.error("Имя пользователя и логин являются пустыми");
-            throw new ValidateException("Имя пользователя и логин являются пустыми");
-        }
-        if (user.getLogin() == null || user.getLogin().contains(" ")) {
-            log.error("Неправильный логин пользователя!");
-            throw new ValidateException("Неправильный логин пользователя!");
-        }
-        if (user.getEmail() == null || !user.getEmail().contains("@")) {
-            log.error("Неправильный email пользователя");
-            throw new ValidateException("Неправильный email пользователя");
-        }
-
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.error("Неправильная дата рождения пользователя");
-            throw new ValidateException("Неправильная дата рождения пользователя");
-        }
-
-
-    }
-
     public void validateUserFriend(Long id) {
         if (id == null && userService.getUserById(id) == null) {
             log.error("Не указаны параметры для удаления из друзей");
@@ -77,19 +54,13 @@ public class UserController {
 
     }
 
-    public void validateAddFriend(Long id, Long friendId) throws ValidateException {
-        if (id == null || friendId == null) {
-            log.error("Не указаны параметры для добавления в друзья");
-            throw new ValidateException("id не может быть равен null");
-        }
-    }
 
     @PutMapping("/{id}/friends/{friendId}")
     public void addFriend(
             @PathVariable Long id,
             @PathVariable Long friendId
     ) {
-        validateAddFriend(id, friendId);
+        userService.validateAddFriend(id, friendId);
         userService.addFriend(id, friendId);
 
     }
